Networking code. Server development. Bug fixes and minor engine changes.
This commit's most major addition is networking code and the corresponding server software.
There is now a standard way to create, send, and process packets in both directions--
client to server and server to client. A basic server has been created which at the moment
can handle commands from the console and is able to send chat down to the client. The
infrastructure is in place to allow for more types of packets in either direction.

Other minor engine changes such as the API on container classes have been handled.
Some minor graphical changes have been added, mainly the new menu icons and partial
chat interface on the client.

Numerous bug fixes which I neglected to document are included in this commit also.

NOTE: The old git repository decided to die on me. Sorry for any strange commit discrepancies.