Modularizes code, adds settings and multiplayer entities.
This commit seperates out KeyHandler from GameCanvas and creates a proper server-
side EntityManager. It also adds the ability to use a configuration file read from
outside the program. The largest change is the overhaul of the entity system and
subsequent change to allow entities and their locations to be synchronized across
clients. The entity system no longer has a concept of a "user", i.e. the setUser()
method. This worked fine while the engine was single-player only, but after the move
to finish networking code, it no longer fit in with the structure of the program.
You now reference entities via a unique identification number. The 0x07 Add Entity
packet has been added and the 0x05 Entity Move packet is implemented.

Minor bug fixes include: 
 - ChatManager now places most recent chat at the bottom of the list.
 - Map now works on headless servers.
 - ResourceManager now initializes on server start. (Needed for entity overhaul.)
 - Made error messages more descriptive and less intrusive.
 - Chat is now logged.