Adds player nameplates, seperates entity drawing from EntityManager.
This commit adds nameplates in which a given player's username is displayed in above
their in-game character. To streamline future development and to match the code to a
more Object-Oriented Paradigm, all code to draw an Entity or any subclass thereof is
now part of the given class, not EntityManager. This will provide more flexibility
with drawing Entities and make the code much more manageable in the future.